pyproject.toml
[tool.poetry]
name = "viz-image-unobfuscate"
version = "0.1.1"
description = "Unobfuscate Viz-style manga images."
authors = ["Kevin Ramdath <krpent@gmail.com>"]
repository = "https://github.com/minormending/viz-image-unobfuscate"
readme = "README.md"
license = "MIT"
keywords = ["viz", "manga", "unobfuscate", "image"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
piexif = "^1.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
nose2 = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
